Add ability to undo/redo actions in form designer
Add an undo and redo button in the Designer.vue template, added text
indicating the number of undos and redos are available. Clicking the
buttons will undo the previous action or redo the action that was
undone. Committing an action when there are redos available will remove
all those redos and set the head of history to the committed action.

We first obtain the original schema when we mount the component, we also
set the original schema when the form is fetched or imported. We stash
each successful action in an array and keep track of where we are in
that history. Undoing and redoing will simply retrieve that old schema
and set the form's schema to that old schema. Being at the head of the
history just means you're at the last committed action, and being at an
index of -1 means you're retrieving the original schema.

There is an issue with this though as storing schemas in an array can
take up a lot of memory. We will need to look into storing changes in
the schema and applying those changes as patches on top of the original
schema.

Resolves: FORMS-7