Add import as a history action
When the user imports a schema now, it gets added as an action on top of
the history. So they can undo that import or redo it.

Needed to modify the lint though and added an ignore pattern for
parameters starting with _ because we need the last parameter in the
formio function when a component is added to determine if an object was
moved or not. This is just a fine grain detail and we can choose not to
differentiate between the two as both add and move component are
considered add component actions by formio.