Improve memory management for history
Converted from storing the entire schema in an index of the history to
only the changes between schemas. Getting a patch is done by obtaining the original schema, freezing it, and then applying patches until you reach the requested patch. When adding an action to the history, we obtain the patch prior to the latest action and compare it to the latest action then store those changes in the history.