Resolve code style
Made changes according to the comments on pull request 283. Found a bug
while going through the explanatino for why the onSchemaChange method
was not simplified. We did not check for the cases when a component was
being removed during an add action so it was adding an action for when a
user cancels/removes a component before it's even added to the form.