Remove unused variable
Removed the max number of actions in history since we are no longer
storing the entire schema, and each action is not too large. We can
store quite a lot of actions in the history without much issue. This may
be an issue only if the user is storing hundreds of actions, but this is
easily remedied by saving the form and refreshing the page.