Add post-deployment module
Modified the vuex store to include a value called builder which formio
builder uses to generate the available components. There is an initial
value for this builder and it can be updated to register plugins at
runtime. In this example, the checkmatrix component is imported from an
external location, then the component is registered, then the form is
re-rendered.