Add patch limit, fix tab bug
When reaching a certain amount of patches on a user's machine, the
patches will stop recording actions. I think this may have something to
do with complexity of a form, and complexity of changes between patches.
There was also a bug where switching between tabs in the tab component
was considered an action, this was because I forgot to change the check
for boolean instead of null for the componentAddedStart.