move builder object in designerOptions to form modules in vuex store
I moved the builder object in designerOptions to form modules in vuex store so that it can be accessible in the admin controls for the info links.
This ensures that changes made to the builder object will immediately reflect in admin controls. E.g. If a component or component group is removed from the builder object, it will not be accessible in admin controls