Fixed router navigation translation
Translations are seemingly broken when used outside of a Vue component and can't be used in a stage like the beforeEnter router guard. To circumnavigate this, I've kept the locale inside of the router guard, but moved the translation into the Alert and Error views. Even if the message is already translated, it will just ignore it.