Vue3 (#1027)
* Update FormViewerMultiUpload.vue

quick fix to missing module

* Update FormViewerMultiUpload.vue

Added lodash and fixed missing error reference

* fix: Search not working with server-side pagination for MySubmissions and Submission Table

fix: Search not working with server-side pagination for MySubmissions and Submission Table

* fix: fix submissions not showing in MySubmissionsTable

* FORMS-852 - CSV option 3 export is missing data (#995)

* fix: differentiate single/multi row fields response to avoid missing fields

* fix: some unit tests add and minor fixes

* fix: test row duplication removed and other minor fix

* fix: unit test

* fix: cover additional case for external source fields with json results

* Fixed router navigation translation

Translations are seemingly broken when used outside of a Vue component and can't be used in a stage like the beforeEnter router guard. To circumnavigate this, I've kept the locale inside of the router guard, but moved the translation into the Alert and Error views. Even if the message is already translated, it will just ignore it.

* Fixed some tests

Error and BCGovAlertBanner test and view/components fixed. Added a prop to check if the message needs to be translated.

* Some fixes and sync with main

BaseNotificationBar now allows for translating an unformatted locale message
SubmissionsTable and MySubmissionsTable synced and search is working

---------

Co-authored-by: Ayobami O. Idowu <ayobami.o.idowu@gmail.com>
Co-authored-by: bcvesalink <127142437+bcvesalink@users.noreply.github.com>