Form access validation and formio
Form.io CSS styling moved away from the component level and into the root style.scss file. This seems to fix our issue but is not what we used to do.
Unable to detect nested radio group in the form.idps for validation and there's no way to manually enforce a validation rules check. So when we detect a change in the idpType, we set the form idps ourselves and then change the userType twice to re-validate the radio group that is the parent.