Vue3 (#1044)
* Update FormViewerMultiUpload.vue

quick fix to missing module

* Update FormViewerMultiUpload.vue

Added lodash and fixed missing error reference

* Fixed router navigation translation

Translations are seemingly broken when used outside of a Vue component and can't be used in a stage like the beforeEnter router guard. To circumnavigate this, I've kept the locale inside of the router guard, but moved the translation into the Alert and Error views. Even if the message is already translated, it will just ignore it.

* Fixed some tests

Error and BCGovAlertBanner test and view/components fixed. Added a prop to check if the message needs to be translated.

* Some fixes and sync with main

BaseNotificationBar now allows for translating an unformatted locale message
SubmissionsTable and MySubmissionsTable synced and search is working

* Form access validation and formio

Form.io CSS styling moved away from the component level and into the root style.scss file. This seems to fix our issue but is not what we used to do.
Unable to detect nested radio group in the form.idps for validation and there's no way to manually enforce a validation rules check. So when we detect a change in the idpType, we set the form idps ourselves and then change the userType twice to re-validate the radio group that is the parent.

* Update form.js

sync with master changes for form store

* Fixed errors

SubmissionsTable and MySubmissionsTable were broken but are now fixed
fixed tests

* Update FormSettings.vue

Passed the disabled prop down to the FormFunctionalitySettings so we can check if the form settings were disabled

* Update ManageFormActions.vue

Delete dialog was removed, added back in