Minor changes for exporting
Sometimes formFields in the frontend are given a empty value which is causing it to throw an error and break exporting because it's trying to map an array that isn't an array.

Returned the export fields to always return the correct export type and not a json error message. This also means we won't get an error message on the front end.