replacing pull_request
we replace all instances of pull_request with pull_request_target. this allows forks to access our secrets/variables in the main repo when creating a pull request. there are security concerns with this as anyone could access our secrets but we've changed a setting in the repo to require all outside contributors to require approval from a collaborator before a workflow can run. there is more manual work but this requires one less pull request when delivering changes to main.