Added tests to files
Added some tests to the changed frontend components. There is a bug where the component will ignore any subsequent calls to a method in a component after the first test. Not sure how to resolve this issue, the current workaround is to create a new test file. This obviously should be fixed, but this workaround will be put in place in the mean time to at least get the bug fixed.