prevents export with no submissions (#1264)
* prevents export with no submissions

resolves: https://bcdevex.atlassian.net/browse/FORMS-857

* Prevent export no subs (#1259)

* Minor changes for exporting

Sometimes formFields in the frontend are given a empty value which is causing it to throw an error and break exporting because it's trying to map an array that isn't an array.

Returned the export fields to always return the correct export type and not a json error message. This also means we won't get an error message on the front end.

* fixed tests

* undoing one test

the test for bcgovheader needed to be undone, for some reason it doesn't use the env variables on github

* fixed file download (#1260)

we had to ignore the conditional for if there was any data in the download.. no submissions means no data. export needs to be fixed again though because the base headers should existl

* Delete app/package-lock.json

Unnecessary change

* Revert "Delete app/package-lock.json"

This reverts commit 114a89f72fb5b0ca0484a6f51d502d113bcd3a43.

* Update package-lock.json

undoing package-lock.json changes

* Added tests to files

Added some tests to the changed frontend components. There is a bug where the component will ignore any subsequent calls to a method in a component after the first test. Not sure how to resolve this issue, the current workaround is to create a new test file. This obviously should be fixed, but this workaround will be put in place in the mean time to at least get the bug fixed.

* Update Download.spec.js

fixed broken test