Forms 1039 view submission real estate (#1313)
* Created more space as a submission reviewer

Moved the status panel and notes panel above the form submission instead of beside it when viewing a submission as a reviewer. Changed the notes panel into a data iterator so that it only shows a set amount of notes before you need to change pages.

* Update NotesPanel.vue

Removed comment block of old code.
Replaced constant with a variable.