Added intervals for updating token
Created a function for updating the token. This function is executed when the keycloak detects the token as expired, as well as every normal 60 seconds, as well as whenever there is an update token failure. Upon failure of updating the token, it will create an interval to keep updating the token, upon success, it will clear that interval.