identity provider code fix
when generating the table data in team management, we used to already map the "user.idp.code" into an identityProvider variable in a user object. then we tried to map it again which would result in an undefined value. removed the previous map that we used to do, so that if we wanted in the future, we could change what is displayed in the identity provider column to be more human readable. i.e., using the display name IDIR or Basic BCeID vs idir or bceid-basic.