Fix getAvailableDates
The transformUtils function getAvailableDates that is supposed to grab available dates for a submission period would not always return the same value, as created tests showed. I've rewritten the functions with better documentation and a refactored input. It should still output the same values but it is unknown as the previous documentation for this feature does not actually specify the inputs/outputs nor were there any tests.