refactor: Forms 1118 bcgov composition api (#1344)
* Update BCGovAlertBanner

BCGovAlertBanner updated to the composition API and tests are fixed, added 100% coverage

* Update BCGovFooter

BCGovFooter updated to composition API. Was getting an infinite recursion issue caused by having a computed value inside of the BCGovFooter so I moved the formSubmitMode value to the parent template "App.vue".

* Update BCGovHeader

BCGovHeader has been updated to the composition API and the computed values were moved to its parent template "App.vue" to fix an infinite recursion issue.

* Update BCGovNavBar

BCGovNavBar has been updated to the composition API. Removed the hideNavBar computed method as it's no different from formSubmitMode. Passed it in just like the BCGovHeader and BCGovFooter because of infinite recursion.

* Update BCGovHeader.spec.js

BCGovHeader test has been fixed, it checks for the app title. The app title is passed to it from App.js
Therefore, App.spec.js has been updated to test for it. Coverage is at 100% for both BCGovHeader and App.

---------

Co-authored-by: Vijaivir Dhaliwal <91633223+vijaivir@users.noreply.github.com>