refactor: Forms 1119 profile settings composition api (#1343)
* Update FormLabelProfile

FormLabelProfile has some test coverage

* Update FormAccessSettings

FormAccessSettings has been updated to the composition API and a test file has been created, giving it 100% coverage

* Update FormFunctionalitySettings

FormFunctionalitySettings has been updated to the composition API and has an added test file with 100% coverage

* Update FormGeneralSettings.vue

FormGeneralSettings has been updated to the composition API

* Update FormSubmissionSettings

FormSubmissionSettings has been updated to the composition API.

* Update FormScheduleSettings

FormScheduleSettings has been updated to the composition API and a test file has been created for it. It is close to 100% coverage, it just needs a few cases for AVAILABLE_DATES and AVAILABLE_PERIOD_OPTIONS

* Fix getAvailableDates

The transformUtils function getAvailableDates that is supposed to grab available dates for a submission period would not always return the same value, as created tests showed. I've rewritten the functions with better documentation and a refactored input. It should still output the same values but it is unknown as the previous documentation for this feature does not actually specify the inputs/outputs nor were there any tests.

---------

Co-authored-by: Vijaivir Dhaliwal <91633223+vijaivir@users.noreply.github.com>