feat: Forms 1077 approver role (#1321)
* Added submission approver role

Added submission approver to TeamManagement page
Added the role as a constant for the front end and updated the test
Added a migration file for the role and its permissions
Added to seed data
Added to back end constants
Updated the backend routes to allow for just require submission read for adding notes
Added to cypress

* Update userAccess.spec.js

updated the user access test to include the submission approver

* added submission_review permission

the submission_review permission now determines if a user can get/add notes or get/update the status of a submission. it has been added to the owner and submission_reviewer, submission_approver roles.

* hid the edit button if no update permission

the edit button for a submission is now hidden if you do not have the update permission

* added roles to idps

roles have been added to identity providers using the decoupled identity provider logic.

* updated approver role

approver role has been added to the identity_provider table

* TeamManagement and test fixes

TeamManagement was not updated to use the latest vuetify. AddTeamMember test is fixed.

* Update service.js

Added the submission approver as a default role when a form designer creates a new form.

---------

Co-authored-by: Vijaivir Dhaliwal <91633223+vijaivir@users.noreply.github.com>