Login button visible (#1350)
* Update BaseAuthButton.vue

wrong conditional on login button

* Fixed login button showing

login button was showing even though the user was authenticated. removed it and removed from the router all instances of hasLogin because it gets set in the router but never actually used anywhere except in BaseAuthButton. BaseAuthButton can already check if the user is authenticated by the store though.