Forms 1078 token refresh failure (#1329)
* Added intervals for updating token

Created a function for updating the token. This function is executed when the keycloak detects the token as expired, as well as every normal 60 seconds, as well as whenever there is an update token failure. Upon failure of updating the token, it will create an interval to keep updating the token, upon success, it will clear that interval.

* Update main.js

added token clearing if there was an error when updating the token

* Update main.js

removed deprecated success, error promises. replaced it with a then and catch promise.

---------

Co-authored-by: Vijaivir Dhaliwal <91633223+vijaivir@users.noreply.github.com>