Updated usage of vue-i18n
Any translations are now used by importing the useI18n function and then mapping out the t function and in some cases the locale variable.

Something to look into is the test warnings for $vuetify.open not having the open translation.. as well as a better way to handle isRTL however there may not be a better way.