refactor: FORMS-1122 composition api submissions (#1441)
* Update AuditHistory

AuditHistory has been updated to the composition API with full coverage

* Update DeleteSubmission

DeleteSubmission has been updated to the composition API with full coverage.

* Update ManageSubmissionUsers

ManageSubmissionUsers updated to the composition API but no tests done

* Delete ManageSubmissionUsers.spec.js

Empty test

* Update ManageSubmissionUsers

ManageSubmissionUsers is updated to the composition API and coverage tests were created for it to almost 100%

* MySubmissionsActions updated

MySubmissionsActions updated to composition API with full test coverage

* Update MySubmissionsTable

MySubmissionsTable is updated to the composition API with almost full coverage. The only thing missing is checking if the data in a submission contains duplicate field names. A fix for this would break existing functionality, so we're not going to fix that with this PR.

* Update to MySubmissionsTable

removed the filterData debug and fixed the actions translation for english

* Update NotesPanel

NotesPanel has been updated to the composition API with full coverage

* Update StatusPanel

StatusPanel updated to the composition API and coverage added

* Update StatusTable

StatusTable updated to composition API with full coverage

* Update StatusTable.spec.js

quick performance update

* Update UserDuplicateSubmission

UserDuplicateSubmission updated to composition API with full coverage

* Update UserSubmission

UserSubmission updated to the composition API with full coverage