PrintOptions Updated
PrintOptions is updated to the composition api and has most of the script covered. it is missing watch variables and the print browser functionality as i didn't look deeply into how we can spy on the browsers print functionality. some functions in PrintOptions is moved to a composable or the transformUtils as it can be reused elsewhere and allowed for easier testing.