refactor: Forms 1123 composition forms (#1460)
* Update ExportSubmissions

ExportSubmissions updated to the composition API and almost all coverage is added.. missing the watch variables

* Update FormSubmission

FormSubmission updated to the composition API with almost full coverage. missing a child component check

* PrintOptions Updated

PrintOptions is updated to the composition api and has most of the script covered. it is missing watch variables and the print browser functionality as i didn't look deeply into how we can spy on the browsers print functionality. some functions in PrintOptions is moved to a composable or the transformUtils as it can be reused elsewhere and allowed for easier testing.

* RequestReceipt Updated

RequestReceipt updated to the composition api with full coverage

* SubmissionsTable updated

SubmissionsTable updated to the composition API but test coverage needs to be added

* SubmissionsTable test coverage

SubmissionsTable has almost full coverage.. just missing some checking for an error because we actually skip it. and another for checking if the lodash debounce works.