Update ProactiveHelp
ProactiveHelp was updated to include the key of the object and not just the simple text, each component has a unique key we can use to identify what component it is. When hovering over a component in the builder sidebar, a tooltip should pop up. Clicking the tooltip will pop up the proactive help.. This needs to be heavily optimized, but this is a proof of concept.