build: FORMS-1301 update aws sdk to v3 (#1496)
* build: FORMS-1301 update aws sdk to v3

Upgrade from the v2 to v3 version of the AWS SDK. The v2 is no longer supported.

* fix: FORMS-1301 rewrite storage service for v3

The upgrade from the v2 to v3 version of the AWS SDK has a lot of breaking changes. Rewrote the service to follow the AWS recommendation to await calls, and using the new style of .send(command) for the calls

* fix: FORMS-1301 fix import order

VSCode doesn't alphabetize when being helpful, so did it manually

* fix: FORMS-1301 moved region to a constant

Moved the region value into a constant and added a comment to mention that it isn't used.