refactor: Forms 1120 composition api designer (#1502)
* Updated some files

FormDisclaimer, FormSettings, FormsTable have been updated to the composition API and tests have been updated with improved coverage

* Update FloatButton.vue

Initial upgrade to composition API. This is not working yet, there appear to be recursive issues. This entire template may need to be refactored..

* FloatButton refactored

FloatButton is completely refactored. Need to fix the tests.

* Update FloatButton.vue

Added in the cypress test attributes

* Update FloatButton

Fixed existing FloatButton tests

* Update FloatButton

Full coverage of the view.

* Update FormDesigner

FormDesigner is updated to the composition API but the CSS completely gets changed when this view is rendered.. Perhaps this entire view needs to be refactored.

* Update FormDesigner

Cleaned up FormDesigner, fixed the issue with the styling being broken. The styling in FloatButton was overriding the parent view.

* Update FloatButton.vue

Fixes the colours for FloatButton buttons that are disabled

* Update FormDesigner

Added some test coverage for the FormDesigner

* Update FormDesigner

many tests cases written, mostly with regards to patch history and formio event handling

* Update FormDesigner

FormDesigner tests are almost at 100% coverage. just missing a test for when the language is changed.

* Update FormViewerActions

FormViewerActions was updated to the composition API and full test coverage added

* Update FormViewer.vue

FormViewer is updated to the composition API. This may need to be refactored as there's so much going on in this file.

* Update FormViewer test coverage

Added many lines of test coverage for FormViewer

* Update FormViewer.spec.js

Added many more lines of code coverage

* Update FormViewer tests

FormViewer tests have more coverage, fixed some tests.