Fixed schedule issue
The issue with the schedule was because formatting it to a specific day would mean it's the start of the day. We needed to set it to the end of the day. Formatting it also would turn a date into a string, and we would lose access to the moment capabilities on it. So it's removed in a lot of places. Added tests for the checkIsFormExpired function.