build: FORMS-1799 node update for security fixes (#1599)
* build: FORMS-1799 update node image

Update the node image to address security vulnerabilities reported by RedHat Advanced Cluster Security.

* build: remove VARIANT argument

While the VARIANT argument is useful in some situations, there's the possibility that the Dockerfile is updated with a newer image version but the argument in the devcontainer.json overrides it. Remove it and just use what's in the Dockerfile.

* refactor: fix sonar complaints about dockerfile

Sonar was complaining that the installed packages were not in alphabetical order. It was also complaining that the install wasn't cleaned afterwards.

* refactor: fix sonar complaints about dockerfile

Sonar was complaining that the multiple RUN commands should be combined. Good point!