build: FORMS-1927 use consistent version of formio.js (#1616)
* build: FORMS-1927 formio.js consistency

Use the same version of formio.js for the components and the frontend. As a side effect this will fix a dompurify security vulnerability in the components.

* build: FORMS-1927 formio.js consistency

Use the same version of formio.js for the components and the frontend. As a side effect this will fix a dompurify security vulnerability in the components.