feat/fix: FORMS-1483 queued file handling (#1515)
* queued file handling

Adds the functionality so that files are not uploaded/deleted until the submitter clicks submit or save draft

* Added notification for upload

When a file is chosen, a warning will pop up. This warning will not autodismiss and must be manually closed. It is also unique so that multiple of the same warning do not show up.

* Translate notification

The notification can now be translated and updated when the translation changes

* file deletion error handling

When a file is deleted from the file upload component and an error occurs, it will now display an error message.

* Update FormViewer.vue

Deleting files also occurs after uploading files instead of skipping the deletions if an error occurs when trying to upload files.

* UX changes

the warning colour across chefs is now changed from the yellowish colour to a more dark orange. the save as draft button is changed from an icon to an outlined button with text. changed translation keys from saveAsADraft to saveAsDraft.

* Update FormViewerActions.vue

fixed lint issue

* route to delete files

created a route that deletes many files instead of just 1, added tests for it.

* Update FormViewer.spec.js

fixing some of the sonarqube issues for the getFormSchema function

* Update FormViewer.spec.js

fixed test errors

* Update FormViewer.spec.js

fixed some tests after merge

* Update FormViewer.spec.js

removed tests that no longer exist inside of FormViewer

* Update filePermissions.js

Replaced the static string with the constant from utils, fixing a permission issue with deletion

* Update FormViewer.vue

Replaced the reverse iteration of the upload files with one that iterates and removes files that were successfully uploaded.

---------

Co-authored-by: jasonchung1871 <jasonchung1871@users.noreply.github.com>