clone working on front end
Clone is now working on the front end.
process:
when mounted, check if the form is being duplicated
if the form is being duplicated, grab all the simplefiles from the form and add them to the queue to be uploaded
when the form is being submitted, if we're not duplicating a form, use the old logic for uploading files. if we are then we clone the files in the queue for file uploads. each one returns a new file id and path, we update the submission with the new file ids and paths.
continue with form submission as usual.