Forms 1075 clone file (#1669)
* clone file route

created a route that should clone a file that exists in the s3 store or in local storage

* clone working on front end

Clone is now working on the front end.
process:
when mounted, check if the form is being duplicated
if the form is being duplicated, grab all the simplefiles from the form and add them to the queue to be uploaded
when the form is being submitted, if we're not duplicating a form, use the old logic for uploading files. if we are then we clone the files in the queue for file uploads. each one returns a new file id and path, we update the submission with the new file ids and paths.
continue with form submission as usual.

* Update FormViewer.spec.js

Fixed the failed test for uploading a file.

* Update FormViewer.vue

updates the correct file id

* Update FormViewer.vue

fixed the base path so it's not always /app/ it will use the path that was defined previously, only replacing the file id.

* refactor and tests

refactored a bit of the new code to be cleaner.
updated tests to ensure that the new code doesn't break old tests.
added in tests to check for the copied submission logic.

* Update controller.spec.js

fixed the call for controller.create to controller.clone