SonarQube issues fix
SonarQube did not like that the editForm was not readonly which all of the other components are.  Fixed the exported default function to have a name, it's deviating from the rest of the components now though.

Created a base class for AxiosService which cdogs, ches, and the css service can use. Prevents code duplication issue from sonarqube.