Updated recordsManagement logic
Added a delete route for a retentionPolicy.

Added additional functionality to the changes for a retention policy. When enabling a retention policy, if there were existing submissions that were soft deleted then it will create scheduled deletions for them. If disabling a retention policy, it now deletes it and deletes scheduled submission deletions. If updating a retention policy and it is still enabled, it will recalculate existing scheduled submissions.